op_extra_tasks/requirements.txt
Flask
requests
openai
google-auth